Split ``sphinx.ext.intersphinx._cli``
