+++ .ruff.toml
@@ -395,7 +395,7 @@ select = [
 "sphinx/environment/adapters/toctree.py" = ["B026"]
 
 # whitelist ``print`` for stdout messages
-"sphinx/ext/intersphinx/__init__.py" = ["T201"]
+"sphinx/ext/intersphinx/_cli.py" = ["T201"]
 
 # whitelist ``print`` for stdout messages
 "sphinx/testing/fixtures.py" = ["T201"]
@@ -504,6 +504,7 @@ exclude = [
     "sphinx/ext/doctest.py",
     "sphinx/ext/autosectionlabel.py",
     "sphinx/ext/intersphinx/__init__.py",
+    "sphinx/ext/intersphinx/_cli.py",
     "sphinx/ext/duration.py",
     "sphinx/ext/imgconverter.py",
     "sphinx/ext/imgmath.py",
